+++ format.sh
@@ -1,20 +1,29 @@
 #!/bin/zsh
 
 # format c code
-echo "Formattting C code..."
-clang-format --style=Google -i --verbose \
+echo "Formattting stm32 C code..."
+clang-format --fallback-style=Google -i --verbose \
   stm32/Src/examples/*.c \
   stm32/lib/**/*.h \
   stm32/lib/**/*.c \
   stm32/test/**/*.h \
-  stm32/test/**/*.c \
+  stm32/test/**/*.c
+echo "Done."
+
+echo "Formattting esp32 C++ code..."
+clang-format --fallback-style=Google -i --verbose \
   esp32/**/*.cpp \
-  esp32/**/*.hpp \
+  esp32/**/*.hpp
+echo "Done."
+
+echo "Formatting protobuf code..."
+clang-format --style=Google -i --verbose \
   proto/c/include/transcoder.h \
-  proto/c/src/transcoder.c
+  proto/c/src/transcoder.c \
+  proto/**/*.proto
 echo "Done."
 
 # format python code
 echo "Formattting Python code..."
-black proto/python
+black python
 echo "Done."