+++ stm32/Src/examples/example_sdi12.c
@@ -50,23 +50,11 @@ int main(void) {
 
   /* Initialize all configured peripherals */
   MX_GPIO_Init();
+  MX_DMA_Init();
   MX_USART1_UART_Init();
   MX_USART2_UART_Init();
-  // __HAL_RCC_WAKEUPSTOP_CLK_CONFIG(RCC_STOP_WAKEUPCLOCK_MSI);
-  // Add this back when the rtc is integrated.
-  // Hopefully this will fix the issue.
-  // Then you can remove the nops in SDI12_WakeSensors
-  // MX_RTC_Init();
 
-  /* Ensure that MSI is wake-up system clock */
-  __HAL_RCC_WAKEUPSTOP_CLK_CONFIG(RCC_STOP_WAKEUPCLOCK_MSI);
-
-  /*Initialize timer and RTC*/
-  UTIL_TIMER_Init();
-
-  // HAL_TIM_OnePulse_MspInit(&htim1);
-  // __HAL_RCC_TIM1_CLK_ENABLE();
-  // HAL_TIM_OnePulse_Init(&htim1, TIM_OPMODE_SINGLE);
+  SystemApp_Init();
 
   // User level initialization
 
@@ -90,6 +78,7 @@ int main(void) {
     char buf[32];
     int buf_len = snprintf(buf, sizeof(buf), "0M!");
 
+    // NOLINTNEXTLINE
     if (SDI12GetMeasurment(addr, &measurment_info, buffer, 3000) == HAL_OK) {
       HAL_UART_Transmit(&huart1, (const uint8_t *)success, 7, 100);
       HAL_UART_Transmit(&huart1, buffer, 18, 100);