+++ stm32/Src/lora_app.c
@@ -9,8 +9,7 @@
  *
  * Copyright (c) 2024 STMicroelectronics.
  * All rights reserved.
- *
- * This software is licensed under terms that can be found in the LICENSE file
+ * * This software is licensed under terms that can be found in the LICENSE file
  * in the root directory of this software component.
  * If no LICENSE file comes with this software, it is provided AS-IS.
  *
@@ -42,6 +41,7 @@
 #include "rtc.h"
 #include "sensors.h"
 #include "userConfig.h"
+#include "status_led.h"
 
 #include <time.h>
 /* USER CODE END Includes */
@@ -487,6 +487,8 @@ static void SendTxData(void)
   {
     APP_LOG(TS_OFF, VLEVEL_M, "Could not send request\r\n");
   }
+
+  StatusLedOff();
   /* USER CODE END SendTxData_1 */
 }