+++ stm32/lib/ads/src/ads.c
@@ -1,21 +1,39 @@
 /**
- ******************************************************************************
  * @file    ads.c
  * @author  Stephen Taylor
  * @brief   Soil Power Sensor ADS12 library
- *          This file provides a function to read from the onboard ADC
- *(ADS1219).
  * @date    11/27/2023
  *
- ******************************************************************************
+ * This file provides a function to read from the ADS1219.
+ *
+ * TODO (jmadden173): Improve lib with big bang for regs
+ * TODO (jmadden173): Add offset calibration at startup
  **/
 
-/* Includes ------------------------------------------------------------------*/
 #include "ads.h"
 
 #include <stm32wlxx_hal_gpio.h>
 
 #include "userConfig.h"
+
+/** i2c address */
+static const uint8_t addr = 0x40;
+/** i2c address left shifted one bit for hal i2c funcs */
+static const uint8_t addrls = addr << 1;
+
+/** Reset the device */
+static const uint8_t cmd_reset = 0x06;
+/** Start or restart conversion */
+static const uint8_t cmd_start = 0x08;
+/** Enter power-down mode */
+static const uint8_t cmd_powerdown = 0x02;
+/** Read data */
+static const uint8_t cmd_rdata = 0x10;
+/** Write to register */
+// prev 0x80
+static const uint8_t cmd_rreg = 0x20;
+/** Read from register */
+static const uint8_t cmd_wreg = 0x40;
 
 // #define CALIBRATION
 
@@ -26,11 +44,34 @@ static double current_calibration_m = 0.0;
 static double current_calibration_b = 0.0;
 
 /**
+ * Control register breakdown.
+ *
+ * The implementation for the bit field uses the first value as the LSB. Aka
+ * vref is the LSB and mux is the MSB.
+ *
+ *  7:5 MUX (default)
+ *  4   Gain (default)
+ *  3:2 Data rate (default)
+ *  1   Conversion mode (default)
+ *  0   VREF (External reference 3.3V)
+ */
+typedef union {
+  uint8_t value;
+  struct {
+    uint8_t vref : 1;
+    uint8_t mode : 1;
+    uint8_t dr : 2;
+    uint8_t gain : 1;
+    uint8_t mux : 3;
+  } bits;
+} ConfigReg;
+
+/**
  * @brief GPIO port for adc data ready line
  *
  * @see data_ready_pin
  */
-const GPIO_TypeDef* data_ready_port = GPIOC;
+const GPIO_TypeDef *data_ready_port = GPIOC;
 
 /**
  * @brief GPIO pin for adc data ready line
@@ -38,22 +79,50 @@ const GPIO_TypeDef* data_ready_port = GPIOC;
  */
 const uint16_t data_ready_pin = GPIO_PIN_0;
 
-int HAL_status(HAL_StatusTypeDef ret) {
-  int status;
-  if (ret == HAL_OK) {
-    status = 0;
-  } else if (ret == HAL_ERROR) {
-    status = 1;
-  } else if (ret == HAL_BUSY) {
-    status = 2;
-  } else {
-    status = 3;
-  }
-  return status;
-}
+/** Uart timeout in ms */
+static const unsigned int g_timeout = 5000;
+
+/**
+ * @brief Turn on power to analog circuit
+ *
+ * Has a blocking wait of 1 mS to account for the startup time of OpAmps. This
+ * should be an order of magnitude greater than the startup or turn-on time of
+ * the opamps used in the circuit.
+ *
+ * MAX9944 - N/A
+ * INA296 - 20 uS
+ * THS4532 - 420 nS
+ *
+ * @see data_ready_pin
+ */
+void PowerOn(void);
+
+/**
+ * @brief Turn off power to analog circuit
+ *
+ * @see data_ready_pin
+ */
+void PowerDown(void);
+
+/**
+ * @brief Measure from the adc
+ *
+ * @param mwas Raw measurement
+ *
+ * @return Raw measurement from adc
+ */
+HAL_StatusTypeDef Measure(int32_t *meas);
+
+/**
+ * @brief This function reconfigures the ADS1219 based on the parameter reg_data
+ *
+ * @param reg_data
+ * @return HAL_StatusTypeDef
+ */
+HAL_StatusTypeDef Configure(const ConfigReg reg_data);
 
 HAL_StatusTypeDef ADC_init(void) {
-  const UserConfiguration* cfg = UserConfigGet();
+  const UserConfiguration *cfg = UserConfigGet();
 
   // read calibration values
   voltage_calibration_m = cfg->Voltage_Slope;
@@ -61,191 +130,93 @@ HAL_StatusTypeDef ADC_init(void) {
   current_calibration_m = cfg->Current_Slope;
   current_calibration_b = cfg->Current_Offset;
 
-  uint8_t code = ADS12_RESET_CODE;
-  uint8_t register_data[2] = {0x40, 0x03};
-  HAL_StatusTypeDef ret;
+  HAL_StatusTypeDef ret = HAL_OK;
 
-  // Control register breakdown.
-  //  7:5 MUX (default)
-  //  4   Gain (default)
-  //  3:2 Data rate (default)
-  //  1   Conversion mode (default)
-  //  0   VREF (External reference 3.3V)
-
-  HAL_GPIO_WritePin(
-      GPIOA, GPIO_PIN_9,
-      GPIO_PIN_SET);  // Power down pin has to be set to high before any of the
-                      // analog circuitry can function
-  ret = HAL_I2C_Master_Transmit(&hi2c2, ADS12_WRITE, &code, 1,
-                                HAL_MAX_DELAY);  // Send the reset code
+  // Send the reset code
+  ret = HAL_I2C_Master_Transmit(&hi2c2, addrls, &cmd_reset, 1, g_timeout);
   if (ret != HAL_OK) {
     return ret;
   }
 
-  // Set the control register, leaving everything at default except for the
-  // VREF, which will be set to external reference mode
-  ret = HAL_I2C_Master_Transmit(&hi2c2, ADS12_WRITE, register_data, 2,
-                                HAL_MAX_DELAY);
-  if (ret != HAL_OK) {
-    return ret;
-  }
+  // wait minimum 500 us to reach steady state
+  HAL_Delay(1);
 
-  code = ADS12_START_CODE;
-  ret = HAL_I2C_Master_Transmit(&hi2c2, ADS12_WRITE, &code, 1,
-                                HAL_MAX_DELAY);  // Send a start code
-  if (ret != HAL_OK) {
-    return ret;
-  }
-  HAL_Delay(500);  // Delay to allow ADC start up, not really sure why this is
-                   // neccesary, or why the minimum is 300
   return ret;
 }
 
-HAL_StatusTypeDef ADC_configure(uint8_t reg_data) {
-  uint8_t code = ADS12_RESET_CODE;
-  uint8_t register_data[2] = {0x40, reg_data};
+HAL_StatusTypeDef Configure(const ConfigReg reg_data) {
   HAL_StatusTypeDef ret;
-  uint8_t recx_reg;
-  char reg_string[40];
-
-  // Set the control register, leaving everything at default except for the
-  // VREF, which will be set to external reference mode
-  ret = HAL_I2C_Master_Transmit(&hi2c2, ADS12_WRITE, register_data, 2,
-                                HAL_MAX_DELAY);
-  if (ret != HAL_OK) {
-    return ret;
-  }
-
-  // Send a start code
-  code = ADS12_START_CODE;
-  ret = HAL_I2C_Master_Transmit(&hi2c2, ADS12_WRITE, &code, 1, HAL_MAX_DELAY);
+  const uint8_t i2c_data[2] = {cmd_wreg, reg_data.value};
+  ret = HAL_I2C_Master_Transmit(&hi2c2, addrls, i2c_data, sizeof(i2c_data),
+                                g_timeout);
   return ret;
 }
 
 double ADC_readVoltage(void) {
-  uint8_t code;
-  double reading;
-  HAL_StatusTypeDef ret;
-  uint8_t rx_data[3] = {0x00, 0x00, 0x00};  // Why is this only 3 bytes?
+  HAL_StatusTypeDef ret = HAL_OK;
+  int32_t raw = 0;
+  double meas = 0.0;
 
-  // 0x01 is single shot 0x03 is continuous
-  ret = ADC_configure(0x01);
-  if (ret != HAL_OK) {
-    return -1;
-  }
-  HAL_Delay(60);
-  /*
-  // Modification
-  uint8_t coderreg = 0b00100000; // read from configuration register 0
-  uint8_t rx_datarreg[1] = {0x00}; // rreg only returns 1 byte of data
-  ret = HAL_I2C_Master_Transmit(&hi2c2, ADS12_WRITE, &coderreg, 1,
-  HAL_MAX_DELAY); if (ret != HAL_OK){ return -1;
-  }
-  ret = HAL_I2C_Master_Receive(&hi2c2, ADS12_READ, rx_datarreg, 1, 1000);
-  if (ret != HAL_OK){
-    return -1;
-  }
-  // end of modification
-  */
+  ConfigReg reg_data = {0};
+  reg_data.bits.vref = 1;
 
-  // Wait for the DRDY pin on the ADS12 to go low, this means data is ready
-  while (HAL_GPIO_ReadPin(data_ready_port, data_ready_pin)) {
-  }
-
-  code = ADS12_READ_DATA_CODE;
-  ret = HAL_I2C_Master_Transmit(&hi2c2, ADS12_WRITE, &code, 1, HAL_MAX_DELAY);
-  if (ret != HAL_OK) {
-    return -1;
-  }
-  ret = HAL_I2C_Master_Receive(&hi2c2, ADS12_READ, rx_data, 3, 1000);
+  // 0x21 is single shot and 0x23 is continuos
+  ret = Configure(reg_data);  // configure to read current
   if (ret != HAL_OK) {
     return -1;
   }
 
-  // Combine the 3 bytes into a 24-bit value
-  int32_t temp = ((int32_t)rx_data[0] << 16) | ((int32_t)rx_data[1] << 8) |
-                 ((int32_t)rx_data[2]);
-  // Check if the sign bit (24th bit) is set
-  if (temp & 0x800000) {
-    // Extend the sign to 32 bits
-    temp |= 0xFF000000;
+  ret = Measure(&raw);
+  if (ret != HAL_OK) {
+    return -1;
   }
-  reading = (double)temp;
 
-#ifndef CALIBRATION
-  reading = (voltage_calibration_m * reading) + voltage_calibration_b;
-#endif /* CALIBRATION */
+#ifdef CALIBRATION
+  meas = (double)raw;
+#else
+  meas = (voltage_calibration_m * raw) + voltage_calibration_b;
+  meas /= 1000;
+#endif
 
-  // tmp fix: convert mV to volts
-  reading /= 1000.0;
-
-  return reading;
+  return meas;
 }
 
 double ADC_readCurrent(void) {
-  uint8_t code;
-  double reading;
-  HAL_StatusTypeDef ret;
-  uint8_t rx_data[3] = {0x00, 0x00, 0x00};
+  HAL_StatusTypeDef ret = HAL_OK;
+  int32_t raw = 0;
+  double meas = 0.0;
+
+  ConfigReg reg_data = {0};
+  reg_data.bits.mux = 0b001;
+  reg_data.bits.vref = 1;
 
   // 0x21 is single shot and 0x23 is continuos
-  ret = ADC_configure(0x21);  // configure to read current
-  if (ret != HAL_OK) {
-    return -1;
-  }
-  HAL_Delay(60);
-  /*
-  // Modification
-  uint8_t coderreg = 0b00100000; // read from configuration register 0
-  uint8_t rx_datarreg[1] = {0x00}; // rreg only returns 1 byte of data
-  ret = HAL_I2C_Master_Transmit(&hi2c2, ADS12_WRITE, &coderreg, 1,
-  HAL_MAX_DELAY); if (ret != HAL_OK){ return -1;
-  }
-  ret = HAL_I2C_Master_Receive(&hi2c2, ADS12_READ, rx_datarreg, 1, 1000);
-  if (ret != HAL_OK){
-    return -1;
-  }
-  // end of modification
-  */
-
-  // Wait for the DRDY pin on the ADS12 to go low, this means data is ready
-  while (HAL_GPIO_ReadPin(data_ready_port, data_ready_pin)) {
-  }
-  code = ADS12_READ_DATA_CODE;
-  ret = HAL_I2C_Master_Transmit(&hi2c2, ADS12_WRITE, &code, 1, HAL_MAX_DELAY);
-  if (ret != HAL_OK) {
-    return -1;
-  }
-  ret = HAL_I2C_Master_Receive(&hi2c2, ADS12_READ, rx_data, 3, 1000);
+  ret = Configure(reg_data);  // configure to read current
   if (ret != HAL_OK) {
     return -1;
   }
 
-  // Combine the 3 bytes into a 24-bit value
-  int32_t temp = ((int32_t)rx_data[0] << 16) | ((int32_t)rx_data[1] << 8) |
-                 ((int32_t)rx_data[2]);
-  // Check if the sign bit (24th bit) is set
-  if (temp & 0x800000) {
-    // Extend the sign to 32 bits
-    temp |= 0xFF000000;
+  ret = Measure(&raw);
+  if (ret != HAL_OK) {
+    return -1;
   }
-  reading = (double)temp;
 
-#ifndef CALIBRATION
-  reading = (current_calibration_m * reading) +
-            current_calibration_b;  // Calculated from linear regression
-#endif                              /* CALIBRATION */
+#ifdef CALIBRATION
+  meas = (double)raw;
+#else
+  meas = (current_calibration_m * raw) + current_calibration_b;
+#endif
 
-  return reading;
+  return meas;
 }
 
 HAL_StatusTypeDef probeADS12(void) {
   HAL_StatusTypeDef ret;
-  ret = HAL_I2C_IsDeviceReady(&hi2c2, ADS12_WRITE, 10, 20);
+  ret = HAL_I2C_IsDeviceReady(&hi2c2, addrls, 10, 20);
   return ret;
 }
 
-size_t ADC_measure(uint8_t* data) {
+size_t ADC_measure(uint8_t *data) {
   // get timestamp
   SysTime_t ts = SysTimeGet();
 
@@ -253,7 +224,7 @@ size_t ADC_measure(uint8_t* data) {
   double adc_voltage = ADC_readVoltage();
   double adc_current = ADC_readCurrent();
 
-  const UserConfiguration* cfg = UserConfigGet();
+  const UserConfiguration *cfg = UserConfigGet();
 
   // encode measurement
   size_t data_len = EncodePowerMeasurement(
@@ -261,4 +232,58 @@ size_t ADC_measure(uint8_t* data) {
 
   // return number of bytes in serialized measurement
   return data_len;
+}
+
+void PowerOn(void) {
+  // set high
+  HAL_GPIO_WritePin(POWERDOWN_GPIO_Port, POWERDOWN_Pin, GPIO_PIN_SET);
+  // delay for settling of analog components
+  HAL_Delay(1);
+}
+
+void PowerOff(void) {
+  // set low
+  HAL_GPIO_WritePin(POWERDOWN_GPIO_Port, POWERDOWN_Pin, GPIO_PIN_RESET);
+}
+
+HAL_StatusTypeDef Measure(int32_t *meas) {
+  HAL_StatusTypeDef ret = HAL_OK;
+  uint8_t rx_data[3] = {0x00, 0x00, 0x00};
+
+  PowerOn();
+
+  // start conversion
+  HAL_I2C_Master_Transmit(&hi2c2, addrls, &cmd_start, 1, g_timeout);
+
+  // wait for conversion
+  HAL_Delay(60);
+
+  // Wait for the DRDY pin on the ADS12 to go low, this means data is ready
+  while (HAL_GPIO_ReadPin(data_ready_port, data_ready_pin)) {
+  }
+
+  // send read data command
+  ret = HAL_I2C_Master_Transmit(&hi2c2, addrls, &cmd_rdata, 1, g_timeout);
+  if (ret != HAL_OK) {
+    return -1;
+  }
+
+  // read 3 bytes of data
+  ret = HAL_I2C_Master_Receive(&hi2c2, addrls, rx_data, 3, g_timeout);
+  if (ret != HAL_OK) {
+    return -1;
+  }
+
+  PowerOff();
+
+  // Combine the 3 bytes into a 24-bit value
+  *meas = ((int32_t)rx_data[0] << 16) | ((int32_t)rx_data[1] << 8) |
+          ((int32_t)rx_data[2]);
+  // Check if the sign bit (24th bit) is set
+  if (*meas & 0x800000) {
+    // Extend the sign to 32 bits
+    *meas |= 0xFF000000;
+  }
+
+  return ret;
 }