+++ stm32/test/unity_config.h
@@ -14,7 +14,7 @@
 
 #ifndef NULL
 #ifndef __cplusplus
-#define NULL (void*)0
+#define NULL (void *)0
 #else
 #define NULL 0
 #endif